Tycoon Saving (Zednov's Tycoon Kit)
This project took me about a day and it was my first touch with the Roblox Datastore-System. It work pretty well but in this case it's just more like a rough "Framework". The Player still have to manually claim the Tycoon before it load(!) Keep that in mind. More about how to use this Script in your game look into the description of the Free-Model https://www.roblox.com/library/9028216197/Save-Zednovs-Kit-by-Semiplex